requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
scikit-learn
xgboost
sentence-transformers
shap
pandas
numpy
joblib
kaggle
python-multipart
pdfminer.six
python-docx
python-dotenv
tqdm
aiofiles
